fix: cfg = int

+++ tests/test_numpy.py
@@ -22,7 +22,7 @@ def test_compare_with_reference(dim):
         np.array([[3, 3], [6, 6], [12, 6]])
         if dim == 2
         else np.c_[np.ones((dim,)) * 3, np.ones((dim,)) * 6].T
-    )
+    ).astype(int)
     alpha = 0.3 * rng.uniform(size=1).item()
     r: tuple[float, float, float, float] = tuple(
         np.pi * np.array([1.0, 2.0, 2.0, 4.0]) / 3
@@ -133,7 +133,7 @@ def test_round_trip_absolute(dim):
         np.array([[3, 3], [6, 6], [12, 6]])
         if dim == 2
         else np.c_[np.ones((dim,)) * 3, np.ones((dim,)) * 6].T
-    )
+    ).astype(int)
     alpha = 0.15
     r = tuple(np.pi * np.array([1.0, 2.0, 2.0, 4.0]) / 3)
     winthresh = 1e-5
@@ -183,7 +183,7 @@ def test_round_trip_rel(dim):
         np.array([[3, 3], [6, 6], [12, 6]])
         if dim == 2
         else np.c_[np.ones((dim,)) * 3, np.ones((dim,)) * 6].T
-    )
+    ).astype(int)
     alpha = 0.3 * rng.uniform(size=1)
     r = np.pi * np.array([1.0, 2.0, 2.0, 4.0]) / 3
     winthresh = 10.0 ** (-rng.integers(low=4, high=6, size=1))
